Begins to add select strategy from list functionality
