+++ app/views/students/show.html.erb
@@ -64,7 +64,10 @@
             <%= f.label "Provide a strategy that has helped this student be successful" %>
             <br>
             <div class="strategy-field">
-              <%= f.text_area :strategy, size: "60x10", require: true %>
+              <div class="strategy-examples">
+                <%= f.text_area :strategy, size: "60x10", require: true  %>
+                <%= f.select :strategy, ["Graphic Organizers", "Kinesthetic Learner","Fidget Toy", "Classroom Job", "Team Facilitator"] %>
+              </div>
               <div class="strategy-examples">
                 <p><b>Example</b></p>
                 <p>"Blake can sometimes be jittery and needs assistance staying on tasks. It is helpful for him to have a short list of tasks and he can take a small break in between completion."</p>
@@ -103,6 +106,7 @@
                   <br>
                   <div class="strategy-field">
                     <%= f.text_area :strategy, value:strategy.strategy, size: "60x10", require: true %>
+                    <%= f.select :strategy, ["Graphic Organizers", "Kinesthetic Learner","Fidget Toy", "Classroom Job", "Team Facilitator"] %>
                     <div class="strategy-examples">
                       <p><b>Example</b></p>
                       <p>"Blake can sometimes be jittery and needs assistance staying on tasks. It is helpful for him to have a short list of tasks and he can take a small break in between completion."</p>
